Updated userController functions to validate user type

+++ API/User/UserController.js
@@ -1,8 +1,16 @@
+const messageRouter = require("../Message/MessageRoutes.js");
 const User = require("./UserModel.js");
 
 // Get all users
 exports.getUsers = async (req, res) => {
   try {
+    // Ensure user is admin
+    if (!req.user || !req.user.isAdmin) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized: User is not admin" });
+    }
+
     const users = await User.find();
     res.status(200).send(users);
   } catch (err) {
@@ -26,7 +34,13 @@ exports.getUserById = async (req, res) => {
 // Update user
 exports.updateUserById = async (req, res) => {
   try {
-    const id = req.params.id;
+    const userId = req.params.id;
+
+    // Ensure user is acct owner or admin
+    if (!(req.user.isAdmin || req.user.id === userId || req.user)) {
+      return res.status(403).json({ message: "Unauthorized action" });
+    }
+
     const { email, password, firstName, lastName, birthDate, isAdmin } =
       req.body;
 
@@ -51,6 +65,13 @@ exports.updateUserById = async (req, res) => {
 // Delete User
 exports.deleteUserById = async (req, res) => {
   try {
+    const userId = req.params.id;
+
+    // Ensure user is acct owner or admin
+    if (!(req.user.isAdmin || req.user.id === userId || req.user)) {
+      return res.status(403).json({ message: "Unauthorized action" });
+    }
+
     await User.findByIdAndDelete(req.params.id);
     res.send({ message: "User deleted" });
   } catch (err) {
